Fix for integer data format
Data of unit format in C++ does not exist in R, R automatically
transfer it to numeric. To fix this issue, we define integer vectors in
R to make sure the tstart, tend, astart, and aend are read as integer
type. Talk with Derek who said these fields should have been ints
instead of uints to begin with so that they match the bam file
encoding, we might fix that later.