Avoid Seek on Sequential Access
Often times users will parse files asking for a read at a particular
offset, followed by a call that asks for the next read at a particular
offset.  In the generic case, the buffer is flushed and reloaded with
subsequent system calls on each write.  This change adjust the code so
that if we are already at the correct position to load the next data,
we just go ahead and do that.