Cache Files < 135 kb in Memory
Profiling showed that loading alignments in pbbamr was taking forever
because the code was not caching the data, and was repeatedly going to
disk to load every alignment.  Much of this issue has been avoided by
caching the readers and so avoiding a series of disk calls to verify
the file exists, open it, check for the EOF marker etc.  The remaining
calls occur during parsing of individual BAM files when a random
section of a FASTA file is loaded.  The goal of this commit is to load
those files in memory only once, thereby avoiding a context switch and
a disk read/seek.

This commit tries to fix this by piggy-backing on the already existing
“in-memory” version of the backends htslib provides.  Namely, in the
event a small (< 135 kb) file is detected, the code will read the
entire file into memory and use it as a backend, rather than use a file
handle backend (that is, hFILE_mem instead of hFILE_fd).  To do this I:

* Changed the hopen_fd method to detect small files and return
appropriately (perhaps this makes the method name confusing now though).

* Changed the hFILE_mem structure to indicate if the structure owns the
memory and to free it if so (as it is now responsible for its buffers).

* Rearranged the declaration of the mem_* and fd_* methods as one of
the fd_ methods now calls one of the mem_ methods.  This led to a lot
of chatter.