Fix the indentation
This commit is a follow up of a previous commit that corrects the loadPBI() function. Previously, the correct reference names can only be loaded from BAM files and when loading xml index with multiple bam files, it only loads the numerical index for the reference names. Now the correct reference names are also loaded for the dataset inputs.