Stage loadPBI2, loadExtras
loadPBI2 will replace loadPBI soon.  It handles dataset filters
correctly and with smaller memory overhead.  The API is changed to
separate out the loading of BAM data into a separate routine,
loadExtras.

Formerly, calls to

    loadPBI(..., loadSNR=T,...)

(for example) would be very slow because they would scan the entire
BAM file---even though in the common case we subsequently downsample
the data frame dramatically.

The new approach is to use the following:

    df <- loadPBI2(...)
    df.s <- subsampled(df)
    df.s.withExtras <- loadExtras(df.s, ..., loadSNRT=T,...)

loadExtras, as currently implemented, is seek-oriented, so it works
well for the case where we've subsampled dramatically; it is slower
than the original code if run on an the entire BAM file.  It's not clear
yet if that's a use case we care about.

NOTE that loadPBI2 will be renamed to loadPBI, and the original
loadPBI deleted, once we have some testing confidence.